+++ evennia/contrib/events/scripts.py
@@ -14,9 +14,8 @@ from evennia import logger
 from evennia.utils.create import create_channel
 from evennia.utils.dbserialize import dbserialize
 from evennia.utils.utils import all_from_module, delay
-from evennia.contrib.events.exceptions import InterruptEvent
 from evennia.contrib.events.handler import CallbackHandler
-from evennia.contrib.events.utils import get_next_wait, EVENTS
+from evennia.contrib.events.utils import get_next_wait, EVENTS, InterruptEvent
 
 # Constants
 RE_LINE_ERROR = re.compile(r'^  File "\<string\>", line (\d+)')
